+++ src/3DViewerDocument.ts
@@ -97,9 +97,18 @@ export class Viewer3DDocument implements vscode.CustomDocument {
   private updateColors() {
     if (this.webviewPanel) {
         const config = vscode.workspace.getConfiguration("urdf-editor");
+        // Convert degrees to radians for camera angles
+        const alphaInDegrees = config.get("CameraAlpha", -60);
+        const betaInDegrees = config.get("CameraBeta", 75);
+        const alphaInRadians = (alphaInDegrees * Math.PI) / 180;
+        const betaInRadians = (betaInDegrees * Math.PI) / 180;
+        
         this.webviewPanel.webview.postMessage({ 
             command: 'colors', 
-            cameraRadius: config.get("CameraDistanceToRobot", "1.0"),
+            cameraRadius: config.get("CameraDistanceToRobot", 1),
+            defaultCameraAlpha: alphaInRadians,
+            defaultCameraBeta: betaInRadians,
+            defaultCameraRadius: config.get("CameraDistanceToRobot", 1),
             backgroundColor: config.get("BackgroundColor", "#000000"),
             gridLineColor: config.get("GridMinorColor", "#00FF00"),
             gridMainColor: config.get("GridMainColor", "#001100"),